Move io to program, enable overflow check
